% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representativeness.R
\name{representativeness_config}
\alias{representativeness_config}
\title{Representativeness}
\usage{
representativeness_config(dir, ext = "flt", xy, analysis, dst = NULL,
  job_script = NULL, walltime = NULL, nCPU = 19, mem = NULL,
  logfile = NULL, jobname = NULL, dry_run = TRUE)
}
\arguments{
\item{dir}{(string) Directory where raster grids (transgrids) are located. Will read every file by default.}

\item{ext}{(str) Three letter string denoting the raster file extention (i.e. type of grid - e.g. flt, tif). Default 'flt.'}

\item{xy}{(string) Filepath to a CSV file where locations to sample the grid are located. Assumes x, y, z format. Column names are not important.}

\item{analysis}{(string) The type of analysis to perform. Options are 'min_dis', 'sumsim', 'sumsim_cond.'}

\item{dst}{(string) Directory and filename to save outputs to. If NULL a folder will be created in dir called 'representativeness_<datetime>' to which output files will be written.}

\item{job_script}{(string) Filepath to write jobs script to. Deafult (NULL) will try and use another given arg, else /home/$USER/REPRESENTATIVENESS_<datetime>.sh}

\item{walltime}{(string) Wall time to pass to slurm. If NULL this is guessed.}

\item{nCPU}{(int) Number of CPUs to use (default 19)}

\item{mem}{(int) Memory require to pass to slurm (in GB). If NULL this is guessed.}

\item{logfile}{(str) Filepath to write log file to. Default will be the location from which the analysis is executed (probably /home/$USER/slurm-<JOBID>.log)}

\item{jobname}{(str) Job name to pass to slurm. Default will be default to slurm's choice.}

\item{dry_run}{(bool) If TRUE, job will not be submitted to slurm, but filepath to job script (.sh) returned.}
}
\value{
If dry_run, filepath to job script, else outputs written to dir.
}
\description{
Perform Tom-style representativeness analyses
}
