% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representativeness.R
\name{representativeness_run}
\alias{representativeness_run}
\title{Run representativeness}
\usage{
representativeness_run(job_script)
}
\arguments{
\item{job_script}{(string) Filepath to write jobs script to.}
}
\value{
output from SLURM
}
\description{
Run representativeness using slurm
}
